Trying solving bug

+++ NacloEngine/ModuleScene.cpp
@@ -98,7 +98,10 @@ void ModuleScene::DeleteGameObject(GameObject * go)
 					i--;
 				}
 			}
-			(*it)->parent->DeleteChildren((*it));
+			if ((*it) != root)
+			{
+				(*it)->parent->DeleteChildren((*it));
+			}
 			(*it)->DeleteAllComponents();
 			gameObjects.erase(it);
 			break;
@@ -185,7 +188,7 @@ void ModuleScene::DeleteAllGameObject()
 
 void ModuleScene::DestroyAllGameObjects()
 {
-	gameObjects.clear();
+	DeleteGameObject(root);
 }
 
 bool ModuleScene::CleanUp()
